+++ romancal/tweakreg/tweakreg_step.py
@@ -82,7 +82,6 @@ class TweakRegStep(RomanStep):
 
     def process(self, input):
         use_custom_catalogs = self.use_custom_catalogs
-        catalogs_format = "ascii.ecsv"
 
         if use_custom_catalogs:
             catdict = _parse_catfile(self.catfile)
@@ -98,7 +97,6 @@ class TweakRegStep(RomanStep):
 
         try:
             if use_custom_catalogs and catdict:
-                catalogs_format = self.catalog_format
                 images = datamodels.ModelContainer()
                 if isinstance(input, str):
                     asn_dir = os.path.dirname(input)
@@ -112,24 +110,24 @@ class TweakRegStep(RomanStep):
                             del member["tweakreg_catalog"]
 
                     images.from_asn(input)
-
                 elif is_association(input):
                     images.from_asn(input)
-
                 else:
                     images = datamodels.ModelContainer(input)
                     for im in images:
                         filename = im.meta.filename
                         if filename in catdict:
-                            print(
-                                f"setting {filename}.tweakreg_catalog ="
+                            self.log.info(
+                                f"setting "
+                                f"{filename}.source_detection.tweakreg_catalog_name ="
                                 f" {repr(catdict[filename])}"
                             )
-                            im.meta["tweakreg_catalog"] = catdict[filename]
-
+                            # set catalog name only (no catalog data at this point)
+                            im.meta["source_detection"] = {
+                                "tweakreg_catalog_name": catdict[filename],
+                            }
             else:
                 images = datamodels.ModelContainer(input)
-
         except TypeError as e:
             e.args = (
                 "Input to tweakreg must be a list of DataModels, an "
@@ -157,13 +155,41 @@ class TweakRegStep(RomanStep):
 
         # Build the catalogs for input images
         for i, image_model in enumerate(images):
-            if hasattr(image_model.meta, "tweakreg_catalog"):
-                catalog = Table.read(
-                    image_model.meta.tweakreg_catalog,
-                    format=catalogs_format,
+            if hasattr(image_model.meta, "source_detection"):
+                is_tweakreg_catalog_present = hasattr(
+                    image_model.meta.source_detection, "tweakreg_catalog"
                 )
+                is_tweakreg_catalog_name_present = hasattr(
+                    image_model.meta.source_detection, "tweakreg_catalog_name"
+                )
+                if is_tweakreg_catalog_present:
+                    # read catalog in 4D numpy array format
+                    catalog = Table(
+                        data=image_model.meta.source_detection.tweakreg_catalog.T,
+                        names=("id", "xcentroid", "ycentroid", "flux"),
+                    )
+                elif is_tweakreg_catalog_name_present:
+                    catalog = Table.read(
+                        image_model.meta.source_detection.tweakreg_catalog_name,
+                        format=self.catalog_format,
+                    )
+                else:
+                    raise AttributeError(
+                        "Attribute 'meta.source_detection.tweakreg_catalog' is missing."
+                        "Please either run SourceDetectionStep or provide a"
+                        "custom source catalog."
+                    )
+                # set meta.tweakreg_catalog
+                image_model.meta["tweakreg_catalog"] = catalog
+                # remove 4D numpy array from meta.source_detection
+                if is_tweakreg_catalog_present:
+                    del image_model.meta.source_detection["tweakreg_catalog"]
             else:
-                raise AttributeError("Attribute 'meta.tweakreg_catalog' is missing.")
+                raise AttributeError(
+                    "Attribute 'meta.source_detection' is missing."
+                    "Please either run SourceDetectionStep or provide a"
+                    "custom source catalog."
+                )
 
             for axis in ["x", "y"]:
                 if axis not in catalog.colnames:
@@ -479,7 +505,8 @@ class TweakRegStep(RomanStep):
             else:
                 catalog_format = "ascii.ecsv"
             cat_name = str(catalog)
-            catalog = Table.read(catalog, format=catalog_format)
+            if isinstance(catalog, str):
+                catalog = Table.read(catalog, format=catalog_format)
             catalog.meta["name"] = cat_name
         except OSError:
             self.log.error(f"Cannot read catalog {catalog}")