+++ CHANGES.rst
@@ -1,6 +1,11 @@
 0.10.1 (unreleased)
 ===================
 
+associations
+------------
+
+- Initial association code for GBTDS observations [#661]
+
 
 general
 -------
@@ -16,12 +21,21 @@ general
 
 - Remove explicit dependence on ``stdatamodels``. [#676]
 
+- Drop support for Python 3.8 [#694]
+
 source_detection
 ----------------
 - Added SourceDetection Step to pipeline [#608]
+
 - Added option of fixed random seed for unit tests to avoid intermittent failures from randomness. [#668]
 
 - Fix source detection object instantiation. [#669]
+
+- Small bug fix to ensure that output catalogs are not attached to the file when save_catalogs=False [#684]
+
+astrometric_utils
+-----------------
+- Added option to provide epoch so that the coordinates are corrected by proper motion. [#686]
 
 
 0.10.0 (2023-02-21)